Added location images
Added location images to location pages.

This is not a very clean method. Might have to come back and clean it up, but it works so...

Oh I also added the stock number to the detail page.

Squashed the gitignore:
Realised I should just ignore the folder, instead of each individual file